requirements.txt
flask==2.2.3
Flask-SQLAlchemy==3.0.3
PyMySQL==1.0.2
python-dotenv==1.0.0